+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - 添加 API `/api/broadcast/channels` | Add API `/api/broadcast/channels`
 - 添加 API `/api/broadcast/getChannelInfo` | Add API `/api/broadcast/getChannelInfo`
+- 添加 API `/api/broadcast/postMessage` | Add API `/api/broadcast/postMessage`
 
 ## v0.2.2 / 2023-08-16